Move Modal from components to common
