+++ assets/js/common/Modal/index.js
