+++ assets/js/components/DeregistrationModal/DeregistrationModal.jsx
@@ -4,7 +4,7 @@ import { EOS_CLEANING_SERVICES } from 'eos-icons-react';
 
 import { APPLICATION_TYPE, DATABASE_TYPE } from '@lib/model/sapSystems';
 
-import Modal from '@components/Modal';
+import Modal from '@common/Modal';
 import Button from '@common/Button';
 
 const getContentByType = (type, data) => {
