+++ assets/js/components/Eula/Community.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import Button from '@common/Button';
-import Modal from '@components/Modal';
+import Modal from '@common/Modal';
 
 function Community({ visible, dispatch }) {
   return (
