+++ assets/js/components/Eula/Premium.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import Button from '@common/Button';
-import Modal from '@components/Modal';
+import Modal from '@common/Modal';
 
 function Premium({ visible, dispatch }) {
   return (
